requirements.txt
cryptography==2.9.2
Django==2.2.12
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
Markdown==3.2.2
mysqlclient==1.4.6
PyJWT==1.7.1
wechatpy==1.8.13
Pillow==7.1.2
redis==3.5.2
aliyun-python-sdk-core-v3==2.13.11
django-simpleui==4.0.2
django-cors-headers==3.2.1
qrcode==6.1